requirements.txt
replicate
streamlit-nightly==1.30.1.dev20240129